requirements.txt
nonebot2[fastapi]>=2.1.0
nonebot-adapter-onebot>=2.3.0
nonebot-plugin-apscheduler>=0.3.0
nonebot-plugin-localstore>=0.7.0
nonebot-plugin-alconna>=0.50.0
nonebot-plugin-uninfo>=0.5.0
nonebot-plugin-orm>=0.7.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
toml>=0.10.2
tomli>=2.0.0
httpx>=0.24.0
aiohttp>=3.8.0
# 向量数据库
faiss-cpu>=1.7.4
numpy>=1.24.0
# 异步文件操作
aiofiles>=23.0.0
# 图片处理
pillow>=10.0.0
sketchbook-py==0.2.1
# Web 后台
fastapi>=0.100.0
jinja2>=3.1.0
# 工具
python-dotenv>=1.0.0
